Bugfixes in BadVec, implement LCD support
I'd like for the LCD support to be more of an "optional" thing,
which I guess at the moment if one isn't attached it doesn't
actually care so in a way it is. The larger issue is that the
way that it's implemented uses a generous number of delays to
have a nice fade-in/fade-out effect for the backlight. These
delays are blocking, and I'm not sure if that means it could
miss a message or something. Next up will be to implement some
kind of non-blocking way of controlling the backlight, and
maybe making the LCD message printing itself a little more
generic in the event that someone else somewhere down the line
might have a different use for it.