Remove generated code from repo, provide hints for generating
Maybe this will be revisited in the form of a makefile, but for
housekeeping, automatically generated code has been removed from
the repo. It's just a protobuf file that can be generated by nanopb
so the guide to getting that going should be pretty self explanatory.
If not, check the tasks.json file in the .vscode folder. If you're
using the editor, you can just run the tasks but if not that should
give you the commands you need in order to get the files in a working
order.